+++ R/tree_detection.r
@@ -48,7 +48,7 @@
 #'
 #' plot(las)
 #' with(ttops, rgl::points3d(X, Y, Z, col = "red", size = 5, add = TRUE))
-#' @seealso \link{tree_detection_lmf} \link{tree_detection_ptrees} \link{tree_detection_manual}
+#' @seealso \link{tree_detection_lmf} \link{tree_detection_ptrees} \link{tree_detection_manual} \link{tree_detection_multichm}
 tree_detection = function(x, algorithm, ...)
 {
   if (algorithm == "lmf")
@@ -283,5 +283,124 @@ tree_detection_manual = function(las, detected = NULL, ...)
   rgl::rgl.close()
 
   return(apice[, .(X,Y,Z)])
+}
+
+
+#' Tree top detection based on LMF and multi-CHM
+#'
+#' Find the tree tops positions based on a method described in Eysn et al (2015) (see references) and
+#' propably proposed originaly by Milan Kobal (we did not find original publication). This is a local
+#' maximum filter applied on a multi-canopy height model (see details)
+#'
+#' Describtion adapted from Eysn et al (2015), page 1728, section 3.1.3 Method #3\cr\cr
+#' The method is based on iterative canopy height model generation (CHM) and local maximum filter (LMF)
+#' detection within a moving window for various CHMs. The method works in two general steps, which are
+#' (a) sequential identification of potential trees and (b) filtering of the extracted potential trees.\cr
+#' \itemize{
+#' \item Step (a): From the normalized point cloud, an initial CHM is created by assigning the 95th
+#' height percentile within each raster cell. Based on this CHM, LM are detected and the found
+#' positions and heights are stored in a database. For the next iteration, points in the uppermost layer
+#' of the normalized ALS data are eliminated. The “eliminating” layer is defined as a band below the
+#' current CHM. Based on the filtered data, a new CHM is created, LM are extracted, and the
+#' LM parameters are added to the database. This procedure is carried out sequentially until all points
+#' are removed from the normalized point cloud.
+#' \item Step (b): All detected LM in the database are sorted by decreasing heights.
+#' The highest LM is considered a detected tree. For each following LM, the LM is considered a
+#' detected tree if there is no detected tree within a given 2D distance as well as a given 3D distance.
+#' }
+#' @param las An object of the class LAS
+#' @param res numeric. Resolution of the CHM based on the 95th percentile
+#' @param layer_thickness numeric. The “eliminating” layer is defined as a band of \code{layer_thickness} m
+#' below the current CHM (see details).
+#' @param dist_2d numeric. 2D distance threshold. A local maximum is considered a detected tree
+#' if there is no detected tree within this 2D distance (see details).
+#' @param dist_3d numeric. 3D distance threshold. A local maximum is considered a detected tree
+#' if there is no detected tree within this 3D distance (see details).
+#' @param ... supplementary parameters to be pass to \link{tree_detection_lmf} that is used internally
+#' to find the local maxima.
+#'
+#' @return A data.table with the X, Y, Z coordinates of the tree tops. The tree tops returned are the
+#' true highest points within a given pixel whenever the CHMs where computed with the 95th percentile
+#' of height. Otherwise these maxima are not true maxima and cannot be used in subsequent segmentation
+#' algorithms.
+#' @export
+#' @examples
+#' \dontrun{
+#' LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
+#' las = readLAS(LASfile)
+#'
+#' ttops = tree_detection_multichm(las, 1, ws = 5)
+#'
+#' plot(las)
+#' with(ttops, rgl::points3d(X, Y, Z, col = "red", size = 5, add = TRUE))
+#' }
+#' @references
+#' Eysn, L., Hollaus, M., Lindberg, E., Berger, F., Monnet, J. M., Dalponte, M., … Pfeifer, N. (2015).
+#' A benchmark of lidar-based single tree detection methods using heterogeneous forest data from the
+#' Alpine Space. Forests, 6(5), 1721–1747. https://doi.org/10.3390/f6051721
+tree_detection_multichm = function(las, res, layer_thickness = 0.5, dist_2d = 3, dist_3d = 5, ...)
+{
+  assertive::assert_is_a_number(res)
+  assertive::assert_is_a_number(layer_thickness)
+  assertive::assert_is_a_number(dist_2d)
+  assertive::assert_is_a_number(dist_3d)
+  assertive::assert_all_are_positive(res)
+  assertive::assert_all_are_positive(layer_thickness)
+  assertive::assert_all_are_positive(dist_2d)
+  assertive::assert_all_are_positive(dist_3d)
+
+  . <- X <- Y <- Z <- NULL
+
+  dist_2d = dist_2d^2
+  dist_3d = dist_3d^2
+
+  las_copy = LAS(las@data[, .(X,Y,Z)], las@header)
+  LM = list()
+  chm = as.raster(grid_metrics(las, max(Z), res))
+  i = 1
+
+  while(!is.null(las_copy))
+  {
+    chm95 = grid_metrics(las_copy, quantile(Z, probs = 0.95), res)
+    chm95 = as.raster(chm95)
+    lm  = tree_detection_lmf(chm95, 4)
+    z = chm95[lm]
+    lm  = raster::as.data.frame(lm, xy = TRUE, na.rm = TRUE)
+    data.table::setDT(lm)
+    lm[, z := z]
+    LM[[i]] = lm
+    lasclassify(las_copy, chm95, "chm95")
+    las_copy = lasfilter(las_copy, Z < chm95 - layer_thickness)
+    i = i+1
+  }
+
+  LM = data.table::rbindlist(LM)
+  data.table::setorder(LM, -z)
+
+  detected = LM[1]
+  for(i in 2:nrow(LM))
+  {
+    lm = LM[i]
+    distance2D = (lm$x - detected$x)^2 + (lm$y - detected$y)^2
+    distance3D = distance2D + (lm$z - detected$z)^2
+
+    if (!any(distance2D < dist_2d) & !any(distance3D < dist_3d))
+    {
+      detected = rbind(detected, lm)
+      points(detected)
+    }
+  }
+
+  detected[, layer := NULL]
+  data.table::setnames(detected, names(detected), c("X","Y", "Z"))
+
+  cells = raster::cellFromXY(chm, detected[,1:2])
+  LM = chm
+  LM[] = NA
+  LM[cells] = 1:length(cells)
+  lasclassify(las, LM, "LM")
+  detected = las@data[las@data[, if (!anyNA(.BY)) .I[which.max(Z)], by = LM]$V1]
+  las@data[, LM := NULL]
+  return(detected)
 }